# Conflicts:
#	app/src/main/java/com/cleanseproject/cleanse/activities/UserProfileActivity.java
#	app/src/main/java/com/cleanseproject/cleanse/fragments/HomeFragment.java

+++ app/src/main/java/com/cleanseproject/cleanse/adapters/EventListAdapter.java
@@ -1,7 +1,6 @@
 package com.cleanseproject.cleanse.adapters;
 
 import android.content.Context;
-import android.content.Intent;
 import android.support.annotation.NonNull;
 import android.support.v4.widget.CircularProgressDrawable;
 import android.support.v7.widget.RecyclerView;
@@ -14,18 +13,23 @@ import android.widget.TextView;
 
 import com.bumptech.glide.Glide;
 import com.cleanseproject.cleanse.R;
-import com.cleanseproject.cleanse.activities.EventDetailsActivity;
+import com.cleanseproject.cleanse.activities.HomeActivity;
 import com.cleanseproject.cleanse.dataClasses.Event;
 import com.cleanseproject.cleanse.services.EventManagerService;
 import com.cleanseproject.cleanse.services.ImageManagerService;
 
+import java.text.DateFormat;
 import java.util.ArrayList;
+import java.util.Date;
+import java.util.Locale;
 
 public class EventListAdapter extends RecyclerView.Adapter<EventListAdapter.MyViewHolder> {
 
-    private ArrayList<Event> listaEventos;
+    private final Context context;
+    private final ArrayList<Event> listaEventos;
 
-    public EventListAdapter(ArrayList<Event> listaEventos) {
+    public EventListAdapter(Context context, ArrayList<Event> listaEventos) {
+        this.context = context;
         this.listaEventos = listaEventos;
     }
 
@@ -49,37 +53,37 @@ public class EventListAdapter extends RecyclerView.Adapter<EventListAdapter.MyVi
 
     public class MyViewHolder extends RecyclerView.ViewHolder {
 
-        private EventManagerService eventManagerService;
-        private ImageManagerService imageManagerService;
-        private TextView txtTitulo, txtDistancia;
-        private ImageView ivFoto;
-        private ImageButton btnLike, btnShare;
-        private Context context;
+        private final EventManagerService eventManagerService;
+        private final ImageManagerService imageManagerService;
+        private final TextView txtTitulo;
+        private final TextView txtDistancia;
+        private final ImageView ivFoto;
+        private final ImageButton btnLike;
+        private final Context context;
 
-        public MyViewHolder(View v) {
+        MyViewHolder(View v) {
             super(v);
             txtTitulo = v.findViewById(R.id.tvTitulo);
             txtDistancia = v.findViewById(R.id.txtDistancia);
             btnLike = v.findViewById(R.id.btnLike);
-            btnShare = v.findViewById(R.id.btnShare);
             ivFoto = v.findViewById(R.id.ivEvento);
             context = v.getContext();
             imageManagerService = new ImageManagerService();
             eventManagerService = new EventManagerService();
         }
 
-        public void asignarDatos(Event event) {
+        void asignarDatos(Event event) {
             txtTitulo.setText(event.getName());
             String distancia;
-            if (event.getDistance() >= 1000)
-                distancia = Math.round(event.getDistance() / 1000) + " km";
-            else
-                distancia = Math.round(event.getDistance()) + " m";
+            if (event.getDistance() != -1) {
+                if (event.getDistance() >= 1000)
+                    distancia = Math.round(event.getDistance() / 1000) + " km";
+                else
+                    distancia = Math.round(event.getDistance()) + " m";
+            } else {
+                distancia = formatDate(event.getEventDate());
+            }
             txtDistancia.setText(distancia);
-            if (event.isFavourite())
-                btnLike.setImageResource(R.drawable.corazon_pressed);
-            else
-                btnLike.setImageResource(R.drawable.corazon_transparente);
             imageManagerService.eventImageDownloadUrl(
                     event.getId(),
                     imageUrl -> {
@@ -92,25 +96,42 @@ public class EventListAdapter extends RecyclerView.Adapter<EventListAdapter.MyVi
                                 .placeholder(progressDrawable)
                                 .into(ivFoto);
                     });
-            ivFoto.setOnClickListener(v -> {
-                Intent intent = new Intent(context, EventDetailsActivity.class);
-                intent.putExtra("Evento", event.getId());
-                context.startActivity(intent);
-            });
-            btnLike.setOnClickListener(v -> {
-                if (!event.isFavourite()) {
-                    btnLike.setImageResource(R.drawable.corazon_pressed);
-                    eventManagerService.setEventAsFavourite(event.getId());
-                    event.setFavourite(true);
-                } else {
-                    btnLike.setImageResource(R.drawable.corazon_transparente);
-                    eventManagerService.deleteFavouriteEvent(event.getId());
-                    event.setFavourite(false);
-                }
-            });
+            ivFoto.setOnClickListener(v -> ((HomeActivity) context).showEventDetails(event.getId()));
+            eventManagerService.isUserAdmin(event.getId(),
+                    isAdmin -> {
+                        if (isAdmin) {
+                            btnLike.setImageResource(R.drawable.corazon_pressed);
+                        } else {
+                            if (event.isFavourite())
+                                btnLike.setImageResource(R.drawable.corazon_pressed);
+                            else
+                                btnLike.setImageResource(R.drawable.corazon_transparente);
+                            btnLike.setOnClickListener(v -> {
+                                if (!event.isFavourite()) {
+                                    btnLike.setImageResource(R.drawable.corazon_pressed);
+                                    eventManagerService.setEventAsFavourite(event.getId());
+                                    event.setFavourite(true);
+                                } else {
+                                    btnLike.setImageResource(R.drawable.corazon_transparente);
+                                    eventManagerService.deleteFavouriteEvent(event.getId());
+                                    event.setFavourite(false);
+                                }
+                            });
+                        }
+                    });
         }
+    }
 
-
+    private String formatDate(long time) {
+        Date date = new Date(time);
+        Locale locale;
+        if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.N) {
+            locale = context.getResources().getConfiguration().getLocales().get(0);
+        } else {
+            locale = context.getResources().getConfiguration().locale;
+        }
+        DateFormat dateFormat = DateFormat.getDateInstance(DateFormat.MEDIUM, locale);
+        return dateFormat.format(date);
     }
 
 }
